Split admin/site components into separate folders
