+++ app/js/client.jsx
@@ -8,9 +8,9 @@ var Redirect = Router.Redirect;
 var NotFoundRoute = Router.NotFoundRoute;
 var ChickenApp = require('./chicken_app.jsx');
 var CoffeeApp = require('./coffee_app.jsx');
-var FourOhFour = require('./components/four_oh_four.jsx');
+var FourOhFour = require('./components/site/four_oh_four.jsx');
 var AdminApp = require('./admin_app.jsx');
-var SignIn = require('./components/sign_in.jsx');
+var SignIn = require('./components/admin/sign_in.jsx');
 
 var Client = React.createClass({
   render: function() {
