+++ app/js/components/admin/input.jsx
@@ -7,7 +7,8 @@ module.exports = React.createClass({
   render: function() {
     return (
       <label>{this.props.labelName}
-        <input required={this.props.isRequired} name={this.props.name} placeholder={this.props.placeholder} defaultValue={this.props.default} type="text" ></input>
+        <input required={this.props.isRequired} name={this.props.name} type="text"
+          placeholder={this.props.placeholder} defaultValue={this.props.default}></input>
       </label>
     );
   }
