+++ app/js/components/admin/select.jsx
@@ -6,7 +6,9 @@ var _ = require('lodash');
 module.exports = React.createClass({
   renderOptions: function() {
     return _.map(this.props.options, function(option) {
-      return (<option key={option.value} value={option.value}>{option.name}</option>);
+      return (
+        <option key={option.value} value={option.value}>{option.name}</option>
+      );
     });
   },
