+++ app/js/components/site/banner.jsx
