+++ app/js/components/site/details.jsx
